# Improve performance heap utilization of parser