--- a/src/Dhall/Parser/Expression.hs
+++ b/src/Dhall/Parser/Expression.hs
@@ -758,13 +758,7 @@ missing = do
   return Missing
 
 importType_ :: Parser ImportType
-importType_ = do
-    let predicate c =
-            c == '~' || c == '.' || c == '/' || c == 'h' || c == 'e' || c == 'm'
-
-    _ <- Text.Megaparsec.lookAhead (Text.Megaparsec.satisfy predicate)
-
-    choice [ local, http, env, missing ]
+importType_ = choice [ local, http, env, missing ]
 
 importHashed_ :: Parser ImportHashed
 importHashed_ = do